Query Runner
QueryRunner takes a list of Queries that are initialized in its constructor and provides functions that will call the various functions in the QueryJDBC class which will enable MYSQL queries to be executed. It also has functions to provide the returned data from the Queries. Currently, the event handlers in QueryFrame call these
functions in order to run the Queries.